requirements.txt
psycopg2-binary==2.9.3
sqlalchemy==1.4.41
pydantic==1.10.2
fastapi==0.83.0
loguru==0.6.0
uvicorn==0.18.3

# python -m uvicorn app:app --reload --port 8895